# #1231 Use polygon centroid for 'point' symbol placements