--- a/js/data/bucket/symbol_bucket.js
+++ b/js/data/bucket/symbol_bucket.js
@@ -14,6 +14,8 @@ var clipLine = require('../../symbol/clip_line');
 var util = require('../../util/util');
 var loadGeometry = require('../load_geometry');
 var CollisionFeature = require('../../symbol/collision_feature');
+var findPoleOfInaccessibility = require('../../util/find_pole_of_inaccessibility');
+var classifyRings = require('../../util/classify_rings');
 
 var shapeText = Shaping.shapeText;
 var shapeIcon = Shaping.shapeIcon;
@@ -278,19 +280,27 @@ SymbolBucket.prototype.addFeature = function(lines, shapedText, shapedIcon, feat
         iconAlongLine = layout['icon-rotation-alignment'] === 'map' && layout['symbol-placement'] === 'line',
         mayOverlap = layout['text-allow-overlap'] || layout['icon-allow-overlap'] ||
             layout['text-ignore-placement'] || layout['icon-ignore-placement'],
-        isLine = layout['symbol-placement'] === 'line',
+        symbolPlacement = layout['symbol-placement'],
+        isLine = symbolPlacement === 'line',
         textRepeatDistance = symbolMinDistance / 2;
 
+    var list = null;
     if (isLine) {
-        lines = clipLine(lines, 0, 0, EXTENT, EXTENT);
+        list = clipLine(lines, 0, 0, EXTENT, EXTENT);
+    } else {
+        // Only care about looping through the outer rings
+        list = classifyRings(lines, 0);
     }
 
-    for (var i = 0; i < lines.length; i++) {
-        var line = lines[i];
+    for (var i = 0; i < list.length; i++) {
+        var anchors = null;
+        // At this point it is a list of points for a line or a list of polygon rings
+        var pointsOrRings = list[i];
+        var line = null;
 
         // Calculate the anchor points around which you want to place labels
-        var anchors;
         if (isLine) {
+            line = pointsOrRings;
             anchors = getAnchors(
                 line,
                 symbolMinDistance,
@@ -303,8 +313,12 @@ SymbolBucket.prototype.addFeature = function(lines, shapedText, shapedIcon, feat
                 EXTENT
             );
         } else {
-            anchors = [ new Anchor(line[0].x, line[0].y, 0) ];
+            line = pointsOrRings[0];
+            anchors = this.findPolygonAnchors(pointsOrRings);
         }
+
+
+        // Here line is a list of points that is either the outer ring of a polygon or just a line
 
         // For each potential label, create the placement features used to check for collisions, and the quads use for rendering.
         for (var j = 0, len = anchors.length; j < len; j++) {
@@ -336,6 +350,23 @@ SymbolBucket.prototype.addFeature = function(lines, shapedText, shapedIcon, feat
                 iconBoxScale, iconPadding, iconAlongLine, {zoom: this.zoom}, feature.properties);
         }
     }
+};
+
+SymbolBucket.prototype.findPolygonAnchors = function(polygonRings) {
+
+    var outerRing = polygonRings[0];
+    if (outerRing.length === 0) {
+        return [];
+    } else if (outerRing.length < 3 || !util.isClosedPolygon(outerRing)) {
+        return [ new Anchor(outerRing[0].x, outerRing[0].y, 0) ];
+    }
+
+    var anchors = null;
+    // 16 here represents 2 pixels
+    var poi = findPoleOfInaccessibility(polygonRings, 16);
+    anchors = [ new Anchor(poi.x, poi.y, 0) ];
+
+    return anchors;
 };
 
 SymbolBucket.prototype.anchorIsTooClose = function(text, repeatDistance, anchor) {