--- a/js/util/classify_rings.js
+++ b/js/util/classify_rings.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var quickselect = require('quickselect');
+var calculateSignedArea = require('./util').calculateSignedArea;
 
 // classifies an array of rings into polygons with outer rings and holes
 module.exports = function classifyRings(rings, maxRings) {
@@ -45,14 +46,4 @@ module.exports = function classifyRings(rings, maxRings) {
 
 function compareAreas(a, b) {
     return b.area - a.area;
-}
-
-function calculateSignedArea(ring) {
-    var sum = 0;
-    for (var i = 0, len = ring.length, j = len - 1, p1, p2; i < len; j = i++) {
-        p1 = ring[i];
-        p2 = ring[j];
-        sum += (p2.x - p1.x) * (p1.y + p2.y);
-    }
-    return sum;
 }