@dpieri 2016-06-08
@blanchg thanks for starting this! Lack of this feature has been a huge pain point for us.

In my opinion, [Pole of Inaccessibility](https://en.wikipedia.org/wiki/Pole_of_inaccessibility) is the ideal way to position labels on polygons. I can't speak to the performance implications of doing it in real time compared to a centroid. If they are negligible then I hope you would consider using this method instead.

Here's a Turf [module](https://github.com/Turfjs/turf-polygon-reduce) that implements it, and some more [background](https://github.com/Turfjs/turf/issues/240).