@blanchg 2016-06-08
That looks like a great solution, however that turf module uses turf-buffer which uses jsts to do the actual bufferring, that would increase the size of mapbox significantly.  I will look to see by how much if we only include the polygon offset/buffer code from jsts.