@blanchg 2016-06-10
@dpieri I found a small library that performs the Pole Of Inaccessibility and extended it to support holes: https://github.com/blanchg/LabelPoint, however it appears to interfere with a lot of polygon draws so that on higher levels the polygons stop rendering completely and the +/- zoom buttons never get displayed.  Is there a way to find out why this is occuring (is there some sort of timeout maybe?) without going step by step through the code?