@blanchg 2016-06-12
Thanks, I found that the simplification can produce zero area polygons with multiple points that was causing issues.