@blanchg 2016-06-13
This is ready now.  It continues to use symbol-placment point.

If the ring is closed e.g. first and last point are the same and has an area != 0 then it chooses the Pole of Inaccessibility otherwise it falls back to the first point as it was before.  This works then for points, lines and polygons that have been simplified down to a point/line.

It adds about 1-2kb to the output, fps benchmarks look very similar.

I am positive there are things I can do better so please let me know!  Thanks