@mollymerp 2016-06-13
@blanchg thank you for contributing! 
could you please:
1. rebase your PR to squash your commits and remove commits from other PRs? you can use `git rebase -i origin/master` to do this.
2. add unit tests and/or [render tests](https://github.com/mapbox/mapbox-gl-test-suite)

thanks! 

👓  @lucaswoj @mourner @mollymerp for review