@blanchg 2016-06-14
Thanks for the reviews @mollymerp and @mourner !

I managed to squish the changes and remove the unrelated ones thanks for pointing that out.  I added some unit tests for polygon_poi and I have a render test but I am uncertain how to proceed to get the render test included.  

I have a feature branch on a fork, I assume I do a PR on the test suite that should only be pulled in once this PR has been?