@blanchg 2016-06-14
I found that building polygons in the render test suite data have outer rings that don't match the tile specification (2.1).  They have both positive and negative areas as outer rings.  I also found util/classify_rings that handles both cases.  Will update to use that.