@blanchg 2016-06-14
Updated to use util/classify_rings which now handles polygons and holes of any winding order.  One raster test will fail until it is updated.  I still need some information on how to get the test-suite aligned with this PR.