@mollymerp 2016-06-15
> I still need some information on how to get the test-suite aligned with this PR.

You can see information on how to update the tests [here](https://github.com/mapbox/mapbox-gl-test-suite#writing-new-tests). Once your tests are satisfactory, make a PR in mapbox/mapbox-gl-test-suite repo and then you can update the sha for that commit in the package.json for this PR. 

I don't see the raster tests failing, I just see the `debug/tile` test failing. Any idea why that might be happening?

@mourner I think you should have final review here.