@blanchg 2016-06-15
@mollymerp Thanks for the information.  I have that PR submitted now.

The debug/tile test fails because one of the labels is in a closed polygon so it is now rendered at the center of that polygon, every other test doesn't render labels on closed polygons.

I have added the line style and zoomed to the label that moved slightly in the debug/tile test.
![image](https://cloud.githubusercontent.com/assets/86495/16094689/bd9fbe62-3384-11e6-969b-15d06a81f5d3.png)