@blanchg 2016-06-17
This is now up to date with the head, squashed, unit tests, there is a render test added and updated the failing debug test.  PR in the test-suite and sha pointing to it in package.json.

I think finally it is ready to merge :smiley: :tada: :balloon: