@mourner 2016-06-17
Fantastic! I'll do a final review tomorrow and merge if everything's good.