@mourner 2016-06-29
This got a bit sidetracked but hopefully we will merge this week, followed by a point release next week.