@blanchg 2016-07-06
@mourner Thanks for looking at this again!  Updated and squashed.