@mourner 2016-07-07
Aside from the comments above, I'm also slightly worried about performance — the algorithm used is pretty expensive. E.g. on a sample 4.6k water polygon, it runs for 350ms. Maybe that's fine for an initial implementation, but we should definitely think about a potential follow-up with a faster algorithm.