@blanchg 2016-07-07
@mourner Thanks for the review, I will address those points.

Can I get that water polygon from somewhere to see if I can improve the performance?  When it is tiled on the map the number of points doesn't seem to get too large though I admit to not having tested on a very complex coastline/waterway.