@mourner 2016-07-08
@blanchg yeah, unfortunately sometimes polygons can be quite big, especially water, landcover and other natural sources. You can use some of the fixtures from https://github.com/mapbox/earcut/tree/master/test/fixtures — the one I tested is the first ring of `water-huge.json`.