@mourner 2016-07-08
@blanchg after spending some time on this, I have designed an algorithm (inspired by this one) that is both faster and _guarantees_ global optimum within given precision (unlike the one from the POI paper). On a few test polygons I tried it on, it runs 6-15x faster. I'll push it to NPM as a module and then we can switch to it in this PR.