@mourner 2016-07-08
Pushed my algorithm to https://github.com/mapbox/polylabel repo. It looks very similar to this one, so the PR could be adapted easily. 

The library I intend to publish handles polygons in GeoJSON format though, so instead of using it as an NPM module, it probably makes sense to bring the code here, replacing `[x, y]` format with the `point-geometry` used in GL JS.