@blanchg 2016-07-11
That looks very promising! Did you want me to pull that into mapbox or leave it for you?  It might be a few days before I get a chance to work on it.