@blanchg 2016-08-12
@mourner if this build passes then using your new method has worked and it is ready for another review.