@mourner 2016-08-12
Awesome! Note that additional advantage of this algorithm is that shapes like rectangles get their center as the label point. 

Can you please run the [benchmarks](https://github.com/mapbox/mapbox-gl-js/blob/master/bench/README.md#running-benchmarks) now to make sure we didn't regress anywhere?