@blanchg 2016-08-12
@mourner Thanks very much for the review I have updated as needed.

# Benchmarks Master/HEAD

**buffer:** 1,435 ms
**fps:** 60 fps
**frame-duration:** 8.4 ms, 3% > 16ms
**query-point:** 1.12 ms
**query-box:** 72.96 ms
**geojson-setdata-small:** 14 ms
**geojson-setdata-large:** 206 ms

# Benchmarks this PR

**buffer:** 1,320 ms
**fps:** 59 fps
**frame-duration:** 8.7 ms, 3% > 16ms
**query-point:** 1.03 ms
**query-box:** 76.07 ms
**geojson-setdata-small:** 13 ms
**geojson-setdata-large:** 210 ms