@blanchg 2016-08-12
@mourner all excellent points, hopefully I addressed them.

# Benchmarks Master/HEAD

**buffer:** 1,435 ms
**fps:** 60 fps
**frame-duration:** 8.4 ms, 3% > 16ms
**query-point:** 1.12 ms
**query-box:** 72.96 ms
**geojson-setdata-small:** 14 ms
**geojson-setdata-large:** 206 ms
# Benchmarks PR

**buffer:** 1,350 ms
**fps:** 59 fps
**frame-duration:** 8.3 ms, 4% > 16ms
**query-point:** 1.00 ms
**query-box:** 72.39 ms
**geojson-setdata-small:** 15 ms
**geojson-setdata-large:** 204 ms